Refactor: MySQLInterface, RedisInterface 생성하여 consumer 파일 코드 리펙토링
